requirements.txt
chromadb
pymupdf
pandas
streamlit
requests
numpy
python-dotenv